backend/requirements.txt
aniso8601==9.0.1
anyio==3.6.2
attrs==22.1.0
certifi==2022.9.24
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.0.0
fastapi==0.95.0
filelock==3.8.0
Flask==2.2.3
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.9
packaging==21.3
pipenv==2022.10.12
platformdirs==2.5.2
pluggy==1.0.0
pydantic==1.10.7
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==1.0.0
python-gql==0.2.4
python-multipart==0.0.6
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.26.1
starlette-graphene3==0.6.0
starlette-graphql==0.2.1
tomli==2.0.1
typing-extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
watchfiles==0.19.0
websockets==11.0
Werkzeug==2.2.3
zipp==3.15.0
